DMA improvements
Paging a DMA frame now fetches the entire range as a single contiguous
block. This ensures that it won't become fragmented if accessed in an
odd order.

A new DMA helper contains both the physical and virtual address of the
DMA range, and handles allocating it. This reduces the number of things
a caller needs to keep track of, and opens up the opportunity for
exposing custom buffers, etc, down the road.