Fix semaphore re-use issue that was blocking net traffic
By ignoring the writes, the Box around the semaphore was being
immediately dropped. If this was re-allocated to a new semaphore, and
then the write resolved, the read would appear to be complete with a
zero length.
To fix this, pending writes are stored until completed, so that the
boxed semaphores are not dropped until resolved.